Renaming home page to homePage and changing its structure to show only an ItemListContainer for now
