--- src/pages/home/homePage.scss
@@ -1,28 +1,0 @@
-@import "../../styles/variables";
-
-.home-page {
-    display: grid;
-    grid-template-columns: 100vw;
-    grid-template-rows: 70px minmax(calc(100vh - 70px), auto);
-    grid-template-areas:
-        "header"
-        "content";
-
-    &__header {
-        grid-area: header;
-        background-color: $background-black;
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        justify-content: space-between;
-    }
-
-    &__content {
-        grid-area: content;
-        background-color: $background-blue;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: flex-start;
-    }
-}
