+++ src/pages/homePage/homePage.scss
@@ -1,0 +1,1 @@
+@import "../../styles/variables";
